refactor: mark unidic contract fixtures as mock
Rename the UniDic contract files and fixture exports so they no longer look like a real adapter implementation. This makes the mock nature explicit in file names, comments, test names, and README wording.

Made-with: Cursor
